requirements.txt
# --- Core ---
chromadb>=0.5.4
numpy>=1.26
tqdm>=4.66
yaml

# PDF parsing (pure Python; works on 3.13)
pypdf>=4.2

# Embeddings (no torch): FastEmbed + ONNX Runtime
fastembed>=0.3.4
onnxruntime-silicon>=1.17 ; platform_system == "Darwin"
onnxruntime>=1.17 ; platform_system != "Darwin"

# API / server
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.6

# Optional utils
python-dotenv>=1.0

# Streamlit
streamlit